+++ src/main/java/com/ureca/snac/asset/repository/AssetHistoryRepositoryImpl.java
@@ -19,7 +19,7 @@ import java.time.format.DateTimeParseException;
 import java.util.List;
 
 import static com.ureca.snac.asset.entity.QAssetHistory.assetHistory;
-import static com.ureca.snac.member.QMember.member;
+import static com.ureca.snac.member.entity.QMember.member;
 
 @Slf4j
 @Repository
